requirements.txt
alpaca-py
pandas
python-dotenv
sqlalchemy
psycopg2-binary